bug fix in AI generator

+++ app/src/main/java/buontyhunter/common/PercentageHelper.java
@@ -7,7 +7,7 @@ public abstract class PercentageHelper {
     private static Random random = new Random();
 
     public static boolean match(double percentageToMatch) {
-        var choose = random.nextDouble(100);
+        var choose = random.nextInt(100);
         return choose <= percentageToMatch;
     }
